+++ Web/RunAndHikeStore.Web/Areas/Admin/Controllers/OrderController.cs
@@ -56,21 +56,22 @@
             try
             {
 
-                EditOrderDetailViewModel order = await orderService.GetViewModelForEditByIdAsync(id);
-
-                order.PaymentStatuses = this.orderService.GetPaymentStatuses();
-                order.OrderStatuses = this.orderService.GetOrderStatuses();
-
-                if (order == null)
+                if (await orderService.ExistsById(id))
                 {
-                    return BadRequest();
-                }
+                    EditOrderDetailViewModel order = await orderService.GetViewModelForEditByIdAsync(id);
 
-                return View(order);
+                    order.PaymentStatuses = this.orderService.GetPaymentStatuses();
+                    order.OrderStatuses = this.orderService.GetOrderStatuses();
+
+                    return View(order);
+                }
+                else
+                {
+                    return RedirectToAction("Error404NotFound", "Home", new { area = "" });
+                }
             }
             catch (System.Exception)
             {
-
                 ModelState.AddModelError("", "Something went wrong");
                 return View();
             }
@@ -86,13 +87,21 @@
         {
             try
             {
-                model.PaymentStatuses = this.orderService.GetPaymentStatuses();
-                model.OrderStatuses = this.orderService.GetOrderStatuses();
+                if (await orderService.ExistsById(model.OrderId))
+                {
+                    model.PaymentStatuses = this.orderService.GetPaymentStatuses();
+                    model.OrderStatuses = this.orderService.GetOrderStatuses();
 
-                await orderService.Edit(model);
-                TempData[MessageConstant.SuccessMessage] = "Successfully editted!";
+                    await orderService.Edit(model);
+                    TempData[MessageConstant.SuccessMessage] = "Successfully editted!";
 
-                return RedirectToAction("ManageOrders", "Order");
+                    return RedirectToAction("ManageOrders", "Order");
+                }
+                else
+                {
+                     return RedirectToAction("Error404NotFound", "Home", new { area = "" });
+                }
+
             }
             catch (System.Exception)
             {
@@ -110,9 +119,16 @@
         [HttpGet]
         public async Task<IActionResult> Delete(string id)
         {
-            await orderService.Delete(id);
+            if (await orderService.ExistsById(id))
+            {
+                await orderService.Delete(id);
 
-            return RedirectToAction(nameof(this.ManageOrders));
+                return RedirectToAction(nameof(this.ManageOrders));
+            }
+            else
+            {
+                return RedirectToAction("Error404NotFound", "Home", new { area = "" });
+            }
         }
     }
 }
